minor changes from review
